Shoppingvagnen fungerar + ändrade några namn på flikar.

+++ src/main/java/com/example/demo/ItemController.java
@@ -15,9 +15,7 @@ import org.springframework.web.multipart.MultipartFile;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 import javax.sql.DataSource;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -46,6 +44,7 @@ public class ItemController {
     String itemPage(Model model){
         logger.info("itemPage is running");
         List<Item> items=itemRepository.findAll();
+        items.sort(Comparator.comparing(Item::getName));
         model.addAttribute("items", items);
         /*Item item=new Item("Stolar","Nya stolar som vi vill ge bort pga omplanering av kontor.", "Möbel", "Stockholm",50,"Pick-up","givit.png");
         model.addAttribute("item", item);*/
@@ -87,7 +86,10 @@ public class ItemController {
     String processBuy(@PathVariable int id, @RequestParam String addItem, Model model, HttpSession session, HttpServletRequest currentSession){
         Item buyItem=itemRepository.findById(id);
         System.out.println("Buy item with id: " + id);
-        List<Item> addedItems=new ArrayList<>();
+        List<Item> addedItems= (List<Item>) session.getAttribute("shoppingList");
+        if(addedItems==null){
+            addedItems=new ArrayList<>();
+        }
         addedItems.add(buyItem);
         session.setAttribute("shoppingList",addedItems);
         model.addAttribute("session",session);
@@ -102,8 +104,11 @@ public class ItemController {
         List<Item> addedItems= (List<Item>) session.getAttribute("shoppingList");
         for (Item i : addedItems){
             Item removeFromItems=itemRepository.findById(i.getId());
-            itemRepository.delete(removeFromItems);
-            System.out.println("Item removed from repository:"+ removeFromItems.getName());
+            if (removeFromItems!=null){
+                itemRepository.delete(removeFromItems);
+                System.out.println("Item removed from repository:"+ removeFromItems.getName());
+
+            }
         }
         //session.setAttribute("shoppingList",addedItems);
         //model.addAttribute("session",session);
@@ -113,19 +118,21 @@ public class ItemController {
     @PostMapping("/cart")
     String cartEdits(@RequestParam int remove, Model model, HttpSession session, HttpServletRequest currentSession){
         currentSession.getSession();
+        Item foundItem=null;
         List<Item> addedItems= (List<Item>) session.getAttribute("shoppingList");
-            Item removedFromShoppingList=itemRepository.findById(remove);
-            itemRepository.save(removedFromShoppingList);
-        System.out.println("Item added to repository:");
-        for (Item i:addedItems) {
-            if (i.getId()==removedFromShoppingList.getId()){
-                addedItems.remove(i);
-                System.out.println("removed from cart:" +addedItems.remove(i));
-
+        for (Item item: addedItems){
+            if (item.getId()==remove){
+                foundItem=item;
+                itemRepository.save(item);
             }
         }
-        session.setAttribute("shoppingList",addedItems);
-        model.addAttribute("session",session);
+        if(foundItem!=null){
+            addedItems.remove(foundItem);
+        }
+            //Item removedFromShoppingList=itemRepository.findById(remove);
+        System.out.println("Item added to repository:");
+
+
 
         return "cart";
     }
